Moved client and server main to examples
