+++ pkg/piecestore/rpc/client/client.go
@@ -1,7 +1,7 @@
 // Copyright (C) 2018 Storj Labs, Inc.
 // See LICENSE for copying information.
 
-package api
+package client
 
 import (
 	"errors"
