+++ pkg/piecestore/rpc/server/server.go
@@ -1,7 +1,7 @@
 // Copyright (C) 2018 Storj Labs, Inc.
 // See LICENSE for copying information.
 
-package api
+package server
 
 import (
 	"bytes"
